chore: restructure program files
