+++ web3.js/src/index.ts
@@ -1,25 +1,19 @@
 export * from './account';
-export * from './address-lookup-table-program';
 export * from './blockhash';
 export * from './bpf-loader-deprecated';
 export * from './bpf-loader';
-export * from './compute-budget';
 export * from './connection';
 export * from './epoch-schedule';
-export * from './ed25519-program';
 export * from './fee-calculator';
 export * from './keypair';
 export * from './loader';
 export * from './message';
 export * from './nonce-account';
+export * from './programs';
 export * from './publickey';
-export * from './stake-program';
-export * from './system-program';
-export * from './secp256k1-program';
 export * from './transaction';
 export * from './validator-info';
 export * from './vote-account';
-export * from './vote-program';
 export * from './sysvar';
 export * from './errors';
 export * from './util/borsh-schema';
