+++ web3.js/src/programs/index.ts
@@ -1,0 +1,7 @@
+export * from './address-lookup-table';
+export * from './compute-budget';
+export * from './ed25519';
+export * from './secp256k1';
+export * from './stake';
+export * from './system';
+export * from './vote';
