+++ web3.js/src/programs/secp256k1.ts
@@ -3,10 +3,10 @@ import * as BufferLayout from '@solana/buffer-layout';
 import secp256k1 from 'secp256k1';
 import sha3 from 'js-sha3';
 
-import {PublicKey} from './publickey';
-import {TransactionInstruction} from './transaction';
-import assert from './util/assert';
-import {toBuffer} from './util/to-buffer';
+import {PublicKey} from '../publickey';
+import {TransactionInstruction} from '../transaction';
+import assert from '../util/assert';
+import {toBuffer} from '../util/to-buffer';
 
 const {publicKeyCreate, ecdsaSign} = secp256k1;
