+++ web3.js/src/programs/system.ts
@@ -5,14 +5,14 @@ import {
   decodeData,
   InstructionType,
   IInstructionInputData,
-} from './instruction';
-import * as Layout from './layout';
-import {NONCE_ACCOUNT_LENGTH} from './nonce-account';
-import {PublicKey} from './publickey';
-import {SYSVAR_RECENT_BLOCKHASHES_PUBKEY, SYSVAR_RENT_PUBKEY} from './sysvar';
-import {Transaction, TransactionInstruction} from './transaction';
-import {toBuffer} from './util/to-buffer';
-import {u64} from './util/bigint';
+} from '../instruction';
+import * as Layout from '../layout';
+import {NONCE_ACCOUNT_LENGTH} from '../nonce-account';
+import {PublicKey} from '../publickey';
+import {SYSVAR_RECENT_BLOCKHASHES_PUBKEY, SYSVAR_RENT_PUBKEY} from '../sysvar';
+import {Transaction, TransactionInstruction} from '../transaction';
+import {toBuffer} from '../util/to-buffer';
+import {u64} from '../util/bigint';
 
 /**
  * Create account system transaction params
