+++ web3.js/test/program-tests/compute-budget.test.ts
@@ -9,9 +9,9 @@ import {
   ComputeBudgetProgram,
   ComputeBudgetInstruction,
   sendAndConfirmTransaction,
-} from '../src';
-import {helpers} from './mocks/rpc-http';
-import {url} from './url';
+} from '../../src';
+import {helpers} from '../mocks/rpc-http';
+import {url} from '../url';
 
 use(chaiAsPromised);
