+++ web3.js/test/program-tests/vote.test.ts
@@ -11,9 +11,9 @@ import {
   sendAndConfirmTransaction,
   SystemInstruction,
   Connection,
-} from '../src';
-import {helpers} from './mocks/rpc-http';
-import {url} from './url';
+} from '../../src';
+import {helpers} from '../mocks/rpc-http';
+import {url} from '../url';
 
 use(chaiAsPromised);
