+++ web3.js/test/transaction.test.ts
@@ -7,8 +7,7 @@ import {Connection} from '../src/connection';
 import {Keypair} from '../src/keypair';
 import {PublicKey} from '../src/publickey';
 import {Transaction, TransactionInstruction} from '../src/transaction';
-import {StakeProgram} from '../src/stake-program';
-import {SystemProgram} from '../src/system-program';
+import {StakeProgram, SystemProgram} from '../src/programs';
 import {Message} from '../src/message';
 import invariant from '../src/util/assert';
 import {toBuffer} from '../src/util/to-buffer';
